Move column_assoc.py to the specs module.
