+++ metricflow/instances.py
@@ -8,7 +8,7 @@ from typing import List, TypeVar, Generic, Tuple
 
 from dbt_semantic_interfaces.references import SemanticModelElementReference, MetricModelReference
 from metricflow.aggregation_properties import AggregationState
-from metricflow.column_assoc import ColumnAssociation
+from metricflow.specs.column_assoc import ColumnAssociation
 from dbt_semantic_interfaces.dataclass_serialization import SerializableDataclass
 from metricflow.specs.specs import (
     MetadataSpec,
