+++ metricflow/plan_conversion/dataflow_to_sql.py
@@ -7,7 +7,7 @@ from typing import Generic, List, Optional, Sequence, TypeVar, Union
 from dbt_semantic_interfaces.objects.aggregation_type import AggregationType
 from dbt_semantic_interfaces.references import MetricModelReference
 from metricflow.aggregation_properties import AggregationState
-from metricflow.column_assoc import ColumnAssociation, SingleColumnCorrelationKey
+from metricflow.specs.column_assoc import ColumnAssociation, SingleColumnCorrelationKey
 from metricflow.filters.time_constraint import TimeRangeConstraint
 from metricflow.dag.id_generation import IdGeneratorRegistry
 from metricflow.dataflow.dataflow_plan import (
