+++ metricflow/plan_conversion/spec_transforms.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Sequence, List
 
-from metricflow.column_assoc import ColumnAssociation
+from metricflow.specs.column_assoc import ColumnAssociation
 from metricflow.plan_conversion.select_column_gen import SelectColumnSet
 from metricflow.plan_conversion.sql_expression_builders import make_coalesced_expr
 from metricflow.specs.specs import (
