+++ metricflow/test/dataflow/builder/test_node_data_set.py
@@ -2,7 +2,7 @@ import logging
 
 from dbt_semantic_interfaces.references import SemanticModelElementReference
 from metricflow.aggregation_properties import AggregationState
-from metricflow.column_assoc import ColumnAssociation, SingleColumnCorrelationKey
+from metricflow.specs.column_assoc import ColumnAssociation, SingleColumnCorrelationKey
 from metricflow.dataflow.builder.node_data_set import DataflowPlanNodeOutputDataSetResolver
 from metricflow.dataflow.dataflow_plan import ReadSqlSourceNode, JoinToBaseOutputNode, JoinDescription
 from metricflow.dataflow.sql_table import SqlTable
